backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
supabase>=1.0.0
python-multipart>=0.0.6
pydantic[email]>=2.0.0
httpx>=0.24.0
pillow>=9.5.0
aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0